requirements.txt
playwright>=1.40.0
pandas>=2.2.0
openpyxl>=3.1.2
pyyaml>=6.0.1
jinja2>=3.1.2
pillow>=10.1.0
beautifulsoup4>=4.12.2
lxml>=4.9.3

